+++ R/RcppExports.R
@@ -57,6 +57,14 @@
     .Call(`_RcppCWB__cl_cpos2rbound`, corpus, s_attribute, cpos, registry)
 }
 
+.cl_delete_corpus <- function(corpus, registry) {
+    .Call(`_RcppCWB__cl_delete_corpus`, corpus, registry)
+}
+
+.cl_charset_name <- function(corpus, registry) {
+    .Call(`_RcppCWB__cl_charset_name`, corpus, registry)
+}
+
 .init_cqp <- function() {
     invisible(.Call(`_RcppCWB_init_cqp`))
 }